Parallel K-Means Image Segmentation on GPU
In this project, k-means image segmentation algorithms are designed for
and implemented on both in purely CPU-based structure and in hybrid CPU
and GPU based structure by using CUDA environment. After the
implementation, performance benchmarking of algorithms is conducted and
average computation time gain of 10 is achieved with the usage of GPU in
order to determine the cluster label assignments of the pixels in the
image.